Rivers & Offset
- Added rivers, because why not?
- Fixed offset calculation.

Offset Fix Explanation:
The offset variables used to be constrained to values between 0 and width (or height). This sounds right, but when looking at how tiles are drawn, we see that the first coordinate drawn is offset/3. So if we start at width/3 we would only ever see up to 2*width/3, leaving the last third to never be drawn. Thus, by mapping the constrained value, we can reach our desired range of 0-2*width as the starting point and draw the full range of tiles!